+++ CHANGELOG.md
@@ -68,15 +68,15 @@ has been improved.
 - Set stylix and nvf to use dracula theme by default.
 - Fixed hyprlock conflicting with stylix.
 - Setup nh in a better fashion.
-- Added support for qmk out of the box and added usbutils for lsusb
-  functionality.
+- Added support for qmk out of the box and added usbutils for lsusb functionality.
 - Massive improvement to Hyprland window rules.
 - Removed broken support for Apple Silicon (this may return).
 - Install script improved.
 - Fixed `nix flake check` not working.
 - Added nvidia prime PCI ID variables to host `variables.nix`.
-- Made Ghostty the new default terminal.
 - Added vim keybindings zsh (alt+hjkl).
 - Added (ctrl+hjkl) keybinds for vim movement while in insert mode in neovim.
 - Supports adb out of the box.
 - Ddubs/dwilliam62 helped with the addition of pyprland and scratchpad support. He is now also a maintainer.
+- Added image used by Stylix into the host variables file.
+- Made printing and NFS variables so they can be easily toggled between hosts.